+++ changelog.md
@@ -14,6 +14,7 @@ Versions:
   - Clicking on tile highlights it and showcases zoom-in and information about it
     in the sidebar
 - Added time system, showcased on GUI
+  - Winter is also seen on tiles that can change their appearance as the season comes
 - Added location system, including rendering of locations put on map
   - Made rendering of locations (with ones from premade map being `faction=0`) with
     their info shown when tile is highlighted